Initial cleanup(dev_black branch)
